feat: move shared Spark and Trainer test helpers to kubeflow.common
Move TestCase, SUCCESS, FAILED, DEFAULT_NAMESPACE, TIMEOUT, and RUNTIME
from kubeflow.trainer.test.common and kubeflow.spark.test.common into a
shared kubeflow.common.testing module. Update all 15 test files to import
from the new location. Original modules are kept as thin re-export shims
for backward compatibility.

Closes #700
