+++ kubeflow/spark/test/common.py
@@ -12,31 +12,21 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Shared test utilities and types for Kubeflow Spark tests."""
+"""Shared test utilities and types for Kubeflow Spark tests.
 
-from dataclasses import dataclass, field
-from typing import Any
+Common helpers have moved to kubeflow.common.testing.
+"""
 
-# Common status constants
-SUCCESS = "success"
-FAILED = "failed"
-TIMEOUT = "timeout"
-RUNTIME = "runtime"
-DEFAULT_NAMESPACE = "default"
+from kubeflow.common.testing import (  # noqa: F401
+    DEFAULT_NAMESPACE,
+    FAILED,
+    RUNTIME,
+    SUCCESS,
+    TIMEOUT,
+    TestCase,
+)
 
 # SparkConnect states for mocking
 SPARK_CONNECT_READY = "spark-connect-ready"
 SPARK_CONNECT_PROVISIONING = "spark-connect-provisioning"
 SPARK_CONNECT_FAILED = "spark-connect-failed"
-
-
-@dataclass
-class TestCase:
-    """Test case dataclass for parametrized tests."""
-
-    name: str
-    expected_status: str = SUCCESS
-    config: dict[str, Any] = field(default_factory=dict)
-    expected_output: Any | None = None
-    expected_error: type[Exception] | None = None
-    __test__ = False
