+++ kubeflow/spark/types/types_test.py
@@ -19,6 +19,7 @@ from unittest.mock import patch
 
 import pytest
 
+from kubeflow.common.testing import SUCCESS, TestCase
 from kubeflow.spark.types.types import (
     Driver,
     Executor,
@@ -29,7 +30,6 @@ from kubeflow.spark.types.types import (
     SparkJob,
     SparkJobStatus,
 )
-from kubeflow.trainer.test.common import SUCCESS, TestCase
 
 
 @pytest.mark.parametrize(
