+++ kubeflow/trainer/backends/container/backend_test.py
@@ -28,13 +28,13 @@ from unittest.mock import Mock, patch
 
 import pytest
 
+from kubeflow.common.testing import FAILED, SUCCESS, TestCase
 from kubeflow.trainer.backends.container.adapters.base import (
     BaseContainerClientAdapter,
 )
 from kubeflow.trainer.backends.container.backend import ContainerBackend
 from kubeflow.trainer.backends.container.types import ContainerBackendConfig
 from kubeflow.trainer.constants import constants
-from kubeflow.trainer.test.common import FAILED, SUCCESS, TestCase
 from kubeflow.trainer.types import types
