+++ kubeflow/trainer/backends/container/runtime_loader_test.py
@@ -22,9 +22,9 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
+from kubeflow.common.testing import FAILED, SUCCESS, TestCase
 from kubeflow.trainer.backends.container import runtime_loader
 from kubeflow.trainer.constants import constants
-from kubeflow.trainer.test.common import FAILED, SUCCESS, TestCase
 from kubeflow.trainer.types import types as base_types
 
 # Sample runtime YAML data for testing
