+++ kubeflow/trainer/backends/container/utils_test.py
@@ -15,9 +15,9 @@
 import pytest
 
 from kubeflow.common.constants import UNKNOWN
+from kubeflow.common.testing import FAILED, SUCCESS, TestCase
 from kubeflow.trainer.backends.container import utils as container_utils
 from kubeflow.trainer.constants import constants
-from kubeflow.trainer.test.common import FAILED, SUCCESS, TestCase
 from kubeflow.trainer.types import types
